--- Pipfile
@@ -1,19 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-pyyaml = "==6.0"
-toml = "==0.10.2"
-pydantic = ">=1.8.0"
-aiohttp = ">=3.8.0"
-stealth-paws = {editable = true, path = "."}
-
-[dev-packages]
-pytest = ">=6.0"
-pytest-asyncio = ">=0.15.0"
-black = ">=22.0.0"
-
-[requires]
-python_version = "3.10"
