+++ main.py
@@ -1,0 +1,6 @@
+def main():
+    print("Hello from asyncflow!")
+
+
+if __name__ == "__main__":
+    main()
