+++ pyproject.toml
@@ -1,0 +1,23 @@
+[project]
+name = "asyncflow"
+version = "0.1.0"
+description = "This is a load balancer here"
+readme = "README.md"
+requires-python = ">=3.10"
+
+dependencies = [
+    "pyyaml==6.0",
+    "toml==0.10.2",
+    "pydantic>=1.8.0",
+    "aiohttp>=3.8.0",
+]
+
+[dependency-groups]
+dev = [
+    "pytest>=6.0",
+    "pytest-asyncio>=0.15.0",
+    "black>=22.0.0",
+]
+
+[tool.uv.sources]
+stealth-paws = {path = ".", editable = true}
